Task-3/requirements.txt
fastapi
uvicorn
torch
torchtext
Pillow
torchvision


